xen-tools: Enable aarch64 qemu
Note, this also fixes a bug in the dependecies for xen-tools when vmsep
DISTRO_FEATURE is enabled.

By default (non-vmsep) qemu package brings in all of the binaries,
qemu-system-i386, qemu-system-aarch64, etc etc etc.. So the dependency on
qemu means all of the QEMU binaries are present.

When vmsep is enabled then qemu-system-i386 and qemu-system-aarch64 are
moved into sub-packages.  Due to this just building / installing xen-tools
(when vmsep is enabled) results in a non-functional version, because it now
depends on qemu-system-i386 (or aarch64) package.

This issue is usually hidden because programs like runx require vmsep to
function, but also require the appropriate subpackage.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>
---
Borrowed from https://github.com/Xilinx/meta-virtualization/commits/rel-v2022.2
---